moved album search to its own component
